refactor: rename layout folder, move App component to a separate folder
