+++ star-wars/src/App/App.scss
@@ -1,5 +1,5 @@
-@use './styles/mixins' as mx;
-@use './styles/vars' as v;
+@use '../styles/mixins' as mx;
+@use '../styles/vars' as v;
 
 #root {
   @include mx.wrapper(100%);
