+++ star-wars/src/App/App.tsx
@@ -1,13 +1,13 @@
 import { FC } from 'react';
 
-import { Main } from './pages/main/main';
+import { Main } from '../pages/main/main';
 
 import { Route, RouterProvider, createBrowserRouter, createRoutesFromElements } from 'react-router-dom';
 
 import './App.scss';
-import { NotFoundPage } from './pages/notFound/notFoundPage';
-import { RootLayout } from './Layout/rootLayout';
-import { Details } from './pages/details/details';
+import { NotFoundPage } from '../pages/notFound/notFoundPage';
+import { RootLayout } from '../layout/rootLayout';
+import { Details } from '../pages/details/details';
 
 const router = createBrowserRouter(
   createRoutesFromElements(
@@ -21,11 +21,7 @@ const router = createBrowserRouter(
 );
 
 const App: FC = () => {
-  return (
-    <>
-      <RouterProvider router={router} />
-    </>
-  );
+  return <RouterProvider router={router} />;
 };
 
 export default App;
