+++ star-wars/src/main.tsx
@@ -1,11 +1,11 @@
 import ReactDOM from 'react-dom/client';
-import App from './App.tsx';
-import { ErrorBoundary } from './components/error-boundary/errorBoundary.tsx';
+import App from './App/App.tsx';
+// import { ErrorBoundary } from './components/error-boundary/errorBoundary.tsx';
 
 import './styles/index.scss';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
-  <ErrorBoundary>
-    <App />,
-  </ErrorBoundary>,
+  // <ErrorBoundary>
+  <App />,
+  // </ErrorBoundary>,
 );
